# docs/openapi.yml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v1

servers:
  - url: http://127.0.0.1:8080/api/v1

tags:
  - name: Me
  - name: Auth
  - name: User
  - name: Card
  - name: Transaction

security:
  - accessCookieAuth: [ ]
    refreshCookieAuth: [ ]

paths:
  /auth/register:
    post:
      tags:
        - Auth
      operationId: register
      summary: Registers a new user
      security: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserView'
        '400':
          $ref: '#/components/responses/BadRequestError'
  /auth/login:
    post:
      tags:
        - Auth
      operationId: login
      security: [ ]
      summary: Logs in user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: User successfully logged in
        '404':
          $ref: '#/components/responses/NotFoundError'
  /auth/logout:
    post:
      tags:
        - Auth
      operationId: logout
      summary: Logs out user
      security: [ ]
      responses:
        '200':
          description: User successfully logged out
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users:
    get:
      tags:
        - User
      operationId: getUsers
      summary: Gets all users
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/sizeParam'
        - $ref: '#/components/parameters/sortParam'
        - name: email
          in: query
          required: false
          schema:
            type: string
            format: email
        - name: first_name
          in: query
          required: false
          schema:
            type: string
        - name: last_name
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Get list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserView'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /user/{uuid}:
    get:
      tags:
        - User
      operationId: getUser
      parameters:
        - $ref: '#/components/parameters/uuidParamInPath'
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserView'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      summary: Patches User by UUID
      tags:
        - User
      operationId: patchUser
      parameters:
        - $ref: '#/components/parameters/uuidParamInPath'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPatch'
      responses:
        '200':
          description: User patched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserView'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      summary: Delete User by UUID
      tags:
        - User
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/uuidParamInPath'
      responses:
        '204':
          description: User deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundError'
  /user/me:
    get:
      summary: Get user's info
      tags:
        - Me
      operationId: getMe
      responses:
        '200':
          description: User data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserView'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /user/me/cards:
    get:
      summary: Get user's list of cards
      tags:
        - Me
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/sizeParam'
        - $ref: '#/components/parameters/sortParam'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CardStatus'
          description: Filter by card status
      operationId: getMyCards
      responses:
        '200':
          description: Cards data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardUserView'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /user/me/transactions:
    get:
      summary: Get user's list of transactions
      tags:
        - Me
      operationId: getMyTransactions
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/sizeParam'
        - $ref: '#/components/parameters/sortParam'
        - name: sender_card_id
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/id'
          description: Filter by sender card
        - name: receiver_card_id
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/id'
          description: Filter by receiver card
      responses:
        '200':
          description: List of transactions retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionView'
        '400':
          $ref: '#/components/responses/BadRequestError'

  /cards:
    get:
      summary: Get list of cards
      tags:
        - Card
      operationId: getCards
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/sizeParam'
        - $ref: '#/components/parameters/sortParam'
        - name: user_uuid
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/uuid'
          description: Filter by owner user UUID
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CardStatus'
          description: Filter by card status
      responses:
        '200':
          description: List of cards retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardAdminView'
        '400':
          $ref: '#/components/responses/BadRequestError'

  /user/{uuid}/cards:
    post:
      summary: Create a new card for user
      tags:
        - Card
      operationId: createCard
      parameters:
        - $ref: '#/components/parameters/uuidParamInPath'
          description: User UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardCreate'
      responses:
        '201':
          description: Card created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardAdminView'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /card/{id}/pan:
    get:
      summary: Get decrypted pan of card by ID
      tags:
        - Card
      operationId: getCardPan
      parameters:
        - $ref: '#/components/parameters/idParamInPath'
          description: Card ID
      responses:
        '200':
          description: Card data
          content:
            application/json:
              schema:
                pan:
                  type: string
                  example: 1234567812345678
                  length: 16
        '404':
          $ref: '#/components/responses/NotFoundError'
  /card/{id}:
    get:
      summary: Get card by ID
      tags:
        - Card
      operationId: getCard
      parameters:
        - $ref: '#/components/parameters/idParamInPath'
          description: Card ID
      responses:
        '200':
          description: Card data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardAdminView'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      summary: Patch card by ID
      tags:
        - Card
      operationId: patchCard
      parameters:
        - $ref: '#/components/parameters/idParamInPath'
          description: Card ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardPatch'
      responses:
        '200':
          description: Card patched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardAdminView'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Delete card by ID
      tags:
        - Card
      operationId: deleteCard
      parameters:
        - $ref: '#/components/parameters/idParamInPath'
          description: Card ID
      responses:
        '204':
          description: Card deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundError'

  /transactions:
    get:
      summary: Get list of transactions
      tags:
        - Transaction
      operationId: getTransactions
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/sizeParam'
        - $ref: '#/components/parameters/sortParam'
        - name: sender_card_id
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/id'
          description: Filter by sender card
        - name: receiver_card_id
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/id'
          description: Filter by receiver card
        - name: sender_uuid
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/uuid'
        - name: receiver_uuid
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/uuid'
      responses:
        '200':
          description: List of transactions retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionView'
        '400':
          $ref: '#/components/responses/BadRequestError'

    post:
      summary: Create a new transaction
      tags:
        - Transaction
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreate'
      responses:
        '201':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionView'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /transaction/{id}:
    get:
      summary: Get transaction by ID
      tags:
        - Transaction
      operationId: getTransaction
      parameters:
        - $ref: '#/components/parameters/idParamInPath'
          description: Transaction ID
      responses:
        '200':
          description: Transaction data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionView'
        '404':
          $ref: '#/components/responses/NotFoundError'

components:
  securitySchemes:
    accessCookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    refreshCookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token

  parameters:
    idParamInPath:
      name: id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/id'
    uuidParamInPath:
      name: uuid
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/uuid'
    uuidParamInQuery:
      name: uuid
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/uuid'
    pageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
      description: Page number (starting from 0)
    sizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
      description: Number of elements per page
    sortParam:
      name: sort
      in: query
      schema:
        type: string
        example: created_at:asc
      description: |
        MultiSort by a fields and directions.
        Format: `field1:direction1,field2:direction2,..`.
        Allowed fields: `id`, `created_at`, other (depends on entity fields).
        Allowed directions: `asc`, `desc`.

  responses:
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ForbiddenError:
      description: Invalid authentication credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFoundError:
      description: Element is not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

  schemas:
    ApiError:
      type: object
      description: Standardized error response returned by API when something goes wrong
      required:
        - status
        - error
        - message
        - path
        - timestamp
      properties:
        status:
          type: integer
          format: int32
          description: HTTP status code of the error
          example: 404
        error:
          type: string
          description: HTTP error reason phrase
          example: Not Found
        message:
          type: string
          description: Human-readable error message explaining what went wrong
          example: Requested resource was not found
        path:
          type: string
          description: API endpoint path where the error occurred
          example: /api/v1/auth/register
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred in ISO-8601 format
          example: 2025-10-01T12:00:00.000Z
    TimeData:
      type: object
      required:
        - created_at
        - updated_at
      properties:
        created_at:
          type: string
          format: date-time
          example: '2025-10-18T09:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2025-10-18T09:30:00Z'

    UserBase:
      type: object
      properties:
        email:
          type: string
          format: email
          example: user@example.com
          maxLength: 320
        first_name:
          type: string
          example: John
          maxLength: 128
        last_name:
          type: string
          example: Doe
          maxLength: 128
    UserCreate:
      type: object
      required:
        - email
        - password
        - first_name
        - last_name
      properties:
        first_name:
          $ref: '#/components/schemas/UserBase/properties/first_name'
        last_name:
          $ref: '#/components/schemas/UserBase/properties/last_name'
        email:
          $ref: '#/components/schemas/UserBase/properties/email'
        password:
          type: string
          format: password
          example: StrongPassword123!
          pattern: ^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#$%^&*]).{8,}$
        role:
          $ref: '#/components/schemas/UserRole'
    UserPatch:
      type: object
      properties:
        first_name:
          $ref: '#/components/schemas/UserBase/properties/first_name'
        last_name:
          $ref: '#/components/schemas/UserBase/properties/last_name'
    UserView:
      allOf:
        - $ref: '#/components/schemas/UserBase'
        - type: object
          properties:
            uuid:
              $ref: '#/components/schemas/uuid'
            email:
              $ref: '#/components/schemas/UserBase/properties/email'
            first_name:
              $ref: '#/components/schemas/UserBase/properties/first_name'
            last_name:
              $ref: '#/components/schemas/UserBase/properties/last_name'
            role:
              $ref: '#/components/schemas/UserRole'
            created_at:
              $ref: '#/components/schemas/TimeData/properties/created_at'
            updated_at:
              $ref: '#/components/schemas/TimeData/properties/updated_at'
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          $ref: '#/components/schemas/UserBase/properties/email'
        password:
          type: string
          format: password
          example: StrongPassword123!
          maxLength: 255
          minLength: 8
    UserRole:
      type: string
      enum:
        - ROLE_USER
        - ROLE_ADMIN
    JwtTokens:
      type: object
      required:
        - access
        - refresh
      properties:
        access:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWUsImlhdCI6MTUxNjIzOTAyMn0.KMUFsIDTnFmyG3nMiGM6H9FNFUROf3wh7SmqJp-QV30
        refresh:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWUsImlhdCI6MTUxNjIzOTAyMn0.KMUFsIDTnFmyG3nMiGM6H9FNFUROf3wh7SmqJp-QV30

    CardBase:
      type: object
      properties:
        balance:
          type: number
          
          minimum: 0
          example: 1500.75
          description: Card balance
        status:
          $ref: '#/components/schemas/CardStatus'
        last4:
          type: integer
          minimum: 1000
          maximum: 9999
          example: 1234
          description: Last 4 digits of card number
        expiryMonth:
          type: integer
          minimum: 1
          maximum: 12
          example: 12
        expiryYear:
          type: integer
          minimum: 0
          example: 2028
    CardCreate:
      allOf:
        - $ref: '#/components/schemas/CardBase'
        - type: object
          required:
            - type
            - expiryMonth
            - expiryYear
          properties:
            expiryMonth:
              $ref: '#/components/schemas/CardBase/properties/expiryMonth'
            expiryYear:
              $ref: '#/components/schemas/CardBase/properties/expiryYear'
    CardAdminView:
      allOf:
        - $ref: '#/components/schemas/TimeData'
        - $ref: '#/components/schemas/CardBase'
        - type: object
          properties:
            id:
              $ref: '#/components/schemas/id'
            owner:
              $ref: '#/components/schemas/UserView'
    CardUserView:
      allOf:
        - $ref: '#/components/schemas/TimeData'
        - $ref: '#/components/schemas/CardBase'
        - type: object
          properties:
            id:
              $ref: '#/components/schemas/id'
    CardPatch:
      allOf:
        - $ref: '#/components/schemas/CardBase'
        - type: object
          description: All fields optional for partial update

    CardStatus:
      type: string
      enum:
        - ACTIVE
        - BLOCKED
        - EXPIRED
    
    TransactionBase:
      type: object
      properties:
        from_card_id:
          $ref: '#/components/schemas/id'
        to_card_id:
          $ref: '#/components/schemas/id'
        amount:
          type: number
          
          minimum: 0.01
          example: 250.75
          description: Transfer amount

    TransactionCreate:
      allOf:
        - $ref: '#/components/schemas/TransactionBase'
        - type: object
          required:
            - from_card_id
            - to_card_id
            - amount
          properties:
            from_card_id:
              $ref: '#/components/schemas/TransactionBase/properties/from_card_id'
            to_card_id:
              $ref: '#/components/schemas/TransactionBase/properties/to_card_id'
            amount:
              $ref: '#/components/schemas/TransactionBase/properties/amount'

    TransactionView:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/id'
        from_card_id:
          $ref: '#/components/schemas/id'
        to_card_id:
          $ref: '#/components/schemas/id'
        amount:
          type: number
        createdAt:
          type: string
          format: date-time
          description: Transaction timestamp

    id:
      type: integer
      format: int64
      minimum: 1
      example: 1
    uuid:
      type: string
      format: uuid
      example: a11cb8a8-7f6c-49cf-9e10-192556941ec3